hardware/API/requirements.txt
fastapi
uvicorn
AWSIoTPythonSDK
python-dotenv